+++ studioai/stats/inferential/ttest.py
@@ -11,7 +11,7 @@
 # URL        : https://github.com/john-james-ai/studioai                                           #
 # ------------------------------------------------------------------------------------------------ #
 # Created    : Wednesday June 7th 2023 11:41:00 pm                                                 #
-# Modified   : Monday August 28th 2023 06:14:42 am                                                 #
+# Modified   : Monday August 28th 2023 07:28:21 am                                                 #
 # ------------------------------------------------------------------------------------------------ #
 # License    : MIT License                                                                         #
 # Copyright  : (c) 2023 John James                                                                 #
@@ -22,7 +22,7 @@ import numpy as np
 from scipy import stats
 from dependency_injector.wiring import inject, Provide
 
-from studioai.container import VisualizerContainer
+from studioai.visual.container import VisualizerContainer
 from studioai.visual.seaborn import Visualizer
 from studioai.stats.inferential.profile import StatTestProfile
 from studioai.stats.inferential.base import (